refactor(web): extract ChatLogItem and chat helpers to reduce GameChat.tsx below 500 lines
